only log specific modules instead of all
The purpose of the logger was always to be able to log specific modules
instead of all modules. Similar to how env_logger works by allowing
someone to say RUST_LOG=application=debug but instead make the modules
logged controlled by the developer of the application and not having the
end user specify environment variables.