Replace `atty` dependency with `is-terminal`
The rationale for this change is explained at
https://github.com/cardoe/stderrlog-rs/issues/56. `is-terminal` has a
MSRV of 1.48, so we need to bump our MSRV to match.

I have taken care of updating the documentation and CI workflows with
the new MSRV. A new `Cargo.lock.msrv` file was generated by executing
`cargo +1.48 check` on a workstation with the Rust 1.48 toolchain
installed.